# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movies - API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Gets the list of movies",
                "description": "Returns the list of movies with optional pagination",
                "operationId": "getMoviesList",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of movies per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "No movie(s) found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Store a new movie",
                "description": "Stores a new movie and returns the movie data",
                "operationId": "storeMovie",
                "requestBody": {
                    "description": "Movie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "release_date",
                                    "rating"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "It's a story about a new beginning..."
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    },
                                    "rate": {
                                        "type": "number",
                                        "format": "int",
                                        "example": 5
                                    },
                                    "duration": {
                                        "type": "number",
                                        "format": "int",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "It's a story about a new beginning..."
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "rate": {
                                            "type": "number",
                                            "format": "int",
                                            "example": 5
                                        },
                                        "duration": {
                                            "type": "number",
                                            "format": "int",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/movies/search": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Search movies by name or description",
                "description": "Returns a list of movies that match the search criteria",
                "operationId": "searchMovies",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query for movie name or description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No movies found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get movie by ID",
                "description": "Returns a single movie",
                "operationId": "getMovieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of movie to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "It's a story about a new beginning..."
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "rate": {
                                            "type": "number",
                                            "format": "int",
                                            "example": 8
                                        },
                                        "duration": {
                                            "type": "number",
                                            "format": "int",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update an existing movie",
                "description": "Updates and returns a movie data",
                "operationId": "updateMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of movie that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Movie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "release_date",
                                    "rating"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "It's a story about a new beginning..."
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    },
                                    "rate": {
                                        "type": "number",
                                        "format": "int",
                                        "example": 8
                                    },
                                    "duration": {
                                        "type": "number",
                                        "format": "int",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "It's a story about a new beginning..."
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "rate": {
                                            "type": "number",
                                            "format": "int",
                                            "example": 8
                                        },
                                        "duration": {
                                            "type": "number",
                                            "format": "int",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete a movie",
                "description": "Deletes a movie and returns a success message",
                "operationId": "deleteMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of movie to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Movie deleted successfully"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "500": {
                        "description": "Error deleting movie"
                    }
                }
            }
        }
    }
}